requirements.txt
deepgram-sdk
openai
pyaudio
python-dotenv
websockets
